Update to Calendars table for db v300
See change I515b3e0a for db ugrade details.

+++ src/com/android/exchange/EasSyncService.java
@@ -1122,7 +1122,7 @@ public class EasSyncService extends AbstractSyncService {
         try {
             int status = resp.getStatus();
             if (status == HttpStatus.SC_OK) {
-                int len = (int)resp.getLength();
+                int len = resp.getLength();
                 InputStream is = resp.getInputStream();
                 OutputStream os = null;
                 Uri attachmentUri = AttachmentUtilities.getAttachmentUri(att.mAccountKey, att.mId);
